refactor(matching-management): extract legacy page and add feature flag routing
