+++ app/admin/matching-management/matching-management-v2.tsx
@@ -1,0 +1,5 @@
+'use client';
+
+export default function MatchingManagementV2() {
+  return <div>Matching Management V2 (placeholder)</div>;
+}
